+++ webrtc-sys/build.rs
@@ -139,7 +139,7 @@ fn main() {
             println!("cargo:rustc-link-lib=dylib=va-drm");
 
             builder
-                .file("src/vaapi/vaapi_display.cpp")
+                .file("src/vaapi/vaapi_display_drm.cpp")
                 .file("src/vaapi/vaapi_h264_encoder_wrapper.cpp")
                 .file("src/vaapi/vaapi_encoder_factory.cpp")
                 .file("src/vaapi/h264_encoder_impl.cpp")
