+++ webrtc-sys/src/vaapi/vaapi_encoder_factory.cpp
@@ -3,7 +3,7 @@
 #include <memory>
 
 #include "h264_encoder_impl.h"
-#include "vaapi_display.h"
+#include "vaapi_display_drm.h"
 
 namespace webrtc {
